Move to Docker for development
I wasn't enjoying getting Ruby up and running on a new vagrant box. My
ansible roles had gotten out or date or something.

This was the final non-docker app I was working on anyway - at least
it's consistent now.

There's a docker-compose config for a database and a web process. I'm
still limited by dokku not using compose in production and so I still
only have these two services. The 'worker' is still a cron job under the
web 'dyno'.